requirements.txt
colorama>=0.4.6
fake-useragent>=1.4.0
lxml>=4.9.0
plyer>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
selenium>=4.15.0
webdriver-manager>=4.0.0 